+++ server/seeders/20181030111852-demo-giftcard.js
@@ -22,7 +22,11 @@ module.exports = {
   async up(queryInterface, Sequelize) {
     var data = await getData().catch(err => console.log(err))
     // console.log(data)
-    return queryInterface.bulkInsert('giftcards', _d.map(data, row => _d.pick(row, ['id', 'code', 'customerid', 'createdat', 'updatedat'])), {})
+    return queryInterface.bulkInsert(
+      'giftcards',
+      _d.map(data, row => _d.pick(row, ['id', 'code', 'amount', 'expiry', 'createdat', 'updatedat', 'customerid'])),
+      {}
+    )
   },
 
   down: (queryInterface, Sequelize) => {